% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot.strength}
\alias{boot.strength}
\alias{boot.strength.tbl_df}
\title{Measure arc strength}
\usage{
\method{boot.strength}{tbl_df}(
  data,
  cluster,
  R = 200,
  m = nrow(data),
  algorithm,
  algorithm.args = list(),
  cpdag = TRUE,
  debug = FALSE
)
}
\arguments{
\item{data}{a data frame containing the data the Bayesian network was
    learned from (for \code{arc.strength()}) or that will be used to compute
    the arc strengths (for \code{boot.strength()} and \code{bf.strength()}).}

\item{cluster}{an optional cluster object from package \pkg{parallel}.}

\item{R}{a positive integer, the number of bootstrap replicates.}

\item{m}{a positive integer, the size of each bootstrap replicate.}

\item{algorithm}{a character string, the structure learning algorithm to be
    applied to the bootstrap replicates. See \code{\link[bnlearn]{structure learning}}
    and the documentation of each algorithm for details.}

\item{algorithm.args}{a list of extra arguments to be passed to the learning
    algorithm.}

\item{cpdag}{a boolean value. If \code{TRUE} the (PDAG of) the equivalence
    class is used instead of the network structure itself. It should make it
    easier to identify score-equivalent arcs.}

\item{debug}{a boolean value. If \code{TRUE} a lot of debugging output is
    printed; otherwise the function is completely silent.}
}
\value{
\code{arc.strength()}, \code{boot.strength()}, \code{custom.strength()},
  \code{bf.strength()} and \code{mean()} return an object of class
  \code{bn.strength}; \code{boot.strength()} and \code{custom.strength()} also
  include information about the relative probabilities of arc directions.

  \code{averaged.network()} returns an object of class \code{bn}.

  See \code{\link[bnlearn]{bn.strength class}} and \code{\link[bnlearn]{bn-class}} for details.
}
\description{
Measure the strength of the probabilistic relationships expressed by the arcs
  of a Bayesian network, and use model averaging to build a network containing
  only the significant arcs.
}
\details{
\code{arc.strength()} computes a measure of confidence or strength for each
  arc, while keeping fixed the rest of the network structure.

  If \code{criterion} is a conditional independence test, the strength is a
  p-value (so the lower the value, the stronger the relationship). The
  conditional independence test would be that to drop the arc from the network.
  The only two possible additional arguments are \code{alpha}, which sets the
  significance threshold that is used in \code{strength.plot()}; and \code{B},
  the number of permutations to be generated for each permutation test.

  If \code{criterion} is the label of a score function, the strength is
  measured by the score gain/loss which would be caused by the arc's removal.
  In other words, it is the difference between the score of the network
  in which the arc is not present and the score of the network in which the arc
  is present. Negative values correspond to decreases in the network score
  and positive values correspond to increases in the network score (the stronger
  the relationship, the more negative the difference). There may be additional
  arguments depending on the choice of the score, see \code{\link[bnlearn]{score}} for
  details. The significance threshold is set to \code{0}.

  \code{boot.strength()} estimates the strength of each arc as its empirical
  frequency over a set of networks learned from bootstrap samples. It computes
  the probability of each arc (modulo its direction) and the probabilities of
  each arc's directions conditional on the arc being present in the graph (in
  either direction). The significance threshold is computed automatically from
  the strength estimates.

  \code{bf.strength()} estimates the strength of each arc using Bayes factors
  to overcome the fact that Bayesian posterior scores are not normalised, and
  uses the latter to estimate the probabilities of all possible states of an
  arc given the rest of the network. The significance threshold is set to
  \code{1}.

  \code{custom.strength()} takes a list of networks and estimates arc strength
  in the same way as \cr \code{boot.strength()}.

  Model averaging is supported for objects of class \code{bn.strength} returned
  by \code{\link[bnlearn]{boot.strength}}, \code{\link[bnlearn]{custom.strength}} and
  \code{\link[bnlearn]{bf.strength}}. The returned network contains the arcs whose
  strength is greater than the \code{threshold} attribute of the
  \code{bn.strength} object passed to \code{averaged.network()}.
}
\note{
\code{averaged.network()} typically returns a completely directed graph; an
  arc can be undirected if and only if the probability of each of its directions
  is exactly 0.5. This may happen, for example, if the arc is undirected in all
  the networks being averaged.
}
\examples{
data(learning.test)
res = hc(learning.test)
arc.strength(res, learning.test)

\dontrun{
arcs = boot.strength(learning.test, algorithm = "hc")
arcs[(arcs$strength > 0.85) & (arcs$direction >= 0.5), ]
averaged.network(arcs)

start = random.graph(nodes = names(learning.test), num = 50)
netlist = lapply(start, function(net) {
  hc(learning.test, score = "bde", iss = 10, start = net) })
arcs = custom.strength(netlist, nodes = names(learning.test),
         cpdag = FALSE)
arcs[(arcs$strength > 0.85) & (arcs$direction >= 0.5), ]
modelstring(averaged.network(arcs))
}

bf.strength(res, learning.test, score = "bds", prior = "marginal")
}
\references{
\bold{for model averaging and boostrap strength (confidence):}

  Friedman N, Goldszmidt M, Wyner A (1999). "Data Analysis with Bayesian
	  Networks: A Bootstrap Approach". \emph{Proceedings of the 15th Annual
	  Conference on Uncertainty in Artificial Intelligence}, 196--201.

  \bold{for the computation of the bootstrap strength (confidence) significance
    threshold:}

  Scutari M, Nagarajan R (2013). "On Identifying Significant Edges in Graphical
    Models of Molecular Networks". \emph{Artificial Intelligence in Medicine},
    \strong{57}(3):207--217.
}
\seealso{
\code{\link[bnlearn]{strength.plot}}, \code{\link[bnlearn]{score}},
  \code{\link[bnlearn]{ci.test}}.
}
\author{
Marco Scutari
}
