% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReducedDimWithGraph.R
\name{countEdge}
\alias{countEdge}
\title{countEdge
count edge number based on cluster label and annotation}
\usage{
countEdge(graph, sce, cluster = "label", annotation = "label", coef = FALSE)
}
\description{
countEdge
count edge number based on cluster label and annotation
}
