% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{describe_graph_bn}
\alias{describe_graph_bn}
\title{The functions are adapted from `tidygraph` to show BN relevant information}
\usage{
describe_graph_bn(x)
}
\description{
The functions are adapted from `tidygraph` to show BN relevant information
}
