% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc.R
\name{hc}
\alias{hc}
\title{Score-based structure learning algorithms}
\usage{
hc(...)
}
\arguments{
\item{...}{additional tuning parameters for the network score. See
    \code{\link[bnlearn]{score}} for details.}
}
\value{
A `tbl_graph` object.
}
\description{
Learn the structure of a Bayesian network using a hill-climbing (HC) or a
  Tabu search (TABU) greedy search.
}
\note{
See \code{\link[bnlearn]{structure learning}} for a complete list of structure learning
  algorithms with the respective references.
}
\examples{
library(bnlearn)
hc(as_tibble(alarm))

}
\seealso{
\code{\link[bnlearn]{network scores}}, \link[bnlearn]{constraint-based algorithms},
  \link[bnlearn]{hybrid algorithms}, \link[bnlearn]{local discovery algorithms},
  \link[bnlearn]{alpha.star}.
}
\author{
Marco Scutari
}
