% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sign.R
\name{loadSign}
\alias{loadSign}
\title{loadSign}
\usage{
loadSign(fileName, return_tbl_graph = TRUE)
}
\arguments{
\item{fileName}{the result of SiGN-BN}

\item{return_tbl_graph}{return tbl_graph}
}
\value{
list of edges, nodes, strength, and bn (bnlearn)
}
\description{
Load the output of SiGN-BN (HC+BS)
}
\examples{
loadSign("result.txt")
}
