% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\docType{methods}
\name{tabu,tbl_df-method}
\alias{tabu,tbl_df-method}
\title{hc}
\usage{
\S4method{tabu}{tbl_df}(
  x,
  start = NULL,
  whitelist = NULL,
  blacklist = NULL,
  score = NULL,
  ...,
  debug = FALSE,
  tabu = 10,
  max.tabu = tabu,
  max.iter = Inf,
  maxp = Inf,
  optimized = TRUE
)
}
\description{
Hill-Climbing algorithm using bnlearn() for tibble
}
