% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\docType{class}
\name{tbl_graph-class}
\alias{tbl_graph-class}
\title{The functions are adapted from `tidygraph` to show BN relevant information
on tbl_graph object (under construction)}
\description{
The functions are adapted from `tidygraph` to show BN relevant information
on tbl_graph object (under construction)
}
