% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\docType{methods}
\name{hc}
\alias{hc}
\title{hc}
\usage{
hc(
  x,
  start = NULL,
  whitelist = NULL,
  blacklist = NULL,
  score = NULL,
  ...,
  debug = FALSE,
  restart = 0,
  perturb = 1,
  max.iter = Inf,
  maxp = Inf,
  optimized = TRUE
)
}
\description{
Hill-Climbing algorithm using bnlearn() for tibble
}
