% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReducedDimWithGraph.R
\name{plotReducedDimWithGraph}
\alias{plotReducedDimWithGraph}
\title{plotReducedDimWithGraph}
\usage{
plotReducedDimWithGraph(
  graph,
  sce,
  dimred = "PCA",
  colour_by = "label",
  annotation = "label",
  only_graph = FALSE,
  ...
)
}
\description{
plotReducedDimWithGraph
}
